requirements.txt
sentence-transformers
langchain
langchain-huggingface
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
ipywidgets
-e .
